pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "async-mariadb-connector"
version = "0.1.0"
authors = [
    { name="Chanikya Nelapatla", email="chanikyac01@gmail.com" },
]
description = "A lightweight, production-grade, and asynchronous Python connector for MariaDB."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "aiomysql>=0.2.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.3",
]

[project.urls]
"Homepage" = "https://github.com/chanikkyasaai/async-mariadb-ml"
"Bug Tracker" = "https://github.com/chanikkyasaai/async-mariadb-ml/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "faker>=20.0.0",
    "langchain-core>=0.1.40",
    "sentence-transformers>=2.2.2",
    "numpy>=1.26.0",
    "langsmith>=0.1.0",
    "httpx>=0.25.0",
    "unittest-xml-reporting>=3.2.0",
    "tabulate>=0.9.0",
]
